claude/skills/note-writer/scripts/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html2text>=2020.1.16
python-slugify>=8.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
